fixed userId handling

+++ Server/src/Lobby.h
@@ -29,7 +29,7 @@ public:
 	bool IsEmpty() const { return m_users.empty(); }
 	bool IsFull() const { return m_users.size() >= m_capacity; }
 	int GetCapactity() const { return m_capacity; }
-	int GetUserAmount() const { return m_users.size(); }
+	int GetUserAmount() const { return static_cast<int>(m_users.size()); }
 
 	void Rename(const std::string& inName) { name = inName; }
